Made Statistics not a static class (fixed methods and fields). Added timeElapsed to Statistics, the time that the game starts is recorded in PlayModeFrame and is saved in Statistics. Added score calculations to every obstacle classes except for Hollow Purple Obstacle.

+++ src/main/java/domain/needForSpear/Player.java
@@ -103,7 +103,7 @@ public class Player {
 
             } else {
                 //
-                statistics.setTimeElapsed(Double.parseDouble(loadList.get(0)));
+                statistics.setTimeElapsed(Long.parseLong(loadList.get(0)));
                 statistics.setChances(Integer.parseInt(loadList.get(1)));
                 statistics.setScore(Double.parseDouble(loadList.get(2)));
